Make it a working Parse.com cloud module
This is done in a hackish way, where I replaced all the relative require paths to
"cloud/gocardless-node/...".

When the require module is not available, I `git submodule add` to "cloud/" folder.
If the module is not available, A stub is used. (partial_util_implementation.js)

The returned object is now a Parse.Cloud.HttpResponse instead of the err, response, body
http://parse.com/docs/js/symbols/Parse.Cloud.HTTPResponse.html
You are advised to use the promise version, e.g.

Parse.Cloud.define("getMerchantDetails", function(request, response) {
  var gcConfig = {
      sandbox:
      appId:
      appSecret:
      token:
      merchantId:
  };
  var gocardless = require('cloud/gocardless-node/lib/gocardless.js')(gcConfig);
  gocardless.merchant.getSelf().then(
    function(httpRequest) {
      var a = JSON.parse(httpRequest.text); // => Object
      console.log(JSON.stringify(a));
      response.success(httpRequest);
    }
  );
});

Only confirmed getMerchantDetails is working.